提取各个处理器到单独的类中
