+++ src/background/processors/BaseProcessor.js
@@ -1,0 +1,10 @@
+/**
+ * 处理器基类。
+ * <p>
+ */
+export class BaseProcessor {
+
+  process(opt, kazeLocalData, kazeFun) {
+    console.log('未实现process方法！');
+  }
+}
